Don't wait if game is launched upon boot

+++ RLBotCS/GameState/PlayerMapping.cs
@@ -21,7 +21,7 @@ namespace RLBotCS.GameState
         public PlayerMetadata applyCarSpawn(CarSpawn carSpawn)
         {
             var playerMetadata = getPlayerMetadata(carSpawn);
-            playerIndexToMetadata[(uint)playerMetadata.playerIndex] = playerMetadata;
+            playerIndexToMetadata[playerMetadata.playerIndex] = playerMetadata;
             actorIdToPlayerIndex[playerMetadata.actorId] = (uint)playerMetadata.playerIndex;
             return playerMetadata;
         }
